created services directory, moved all crud functions into services/crud.ts
