+++ src/pages/Project.tsx
@@ -1,10 +1,8 @@
 //import//
-import { FirebaseError } from "firebase/app";
 import { useState, useEffect } from "react";
 import { useParams } from "react-router-dom";
-import { doc, getDoc } from "firebase/firestore";
-import { db } from "../firebase/firebaseConfig";
 import { TProject } from "../types";
+import { getProject } from "../services";
 
 export const Project = () => {
   const params = useParams();
@@ -15,22 +13,8 @@ export const Project = () => {
     paid: false,
   });
 
-  const getProject = async () => {
-    try {
-      const docSnap = await getDoc(doc(db, "projects", `${params.id}`));
-      if (docSnap.exists()) {
-        setProject({ ...docSnap.data() } as TProject);
-      }
-    } catch (error: unknown) {
-      if (error instanceof FirebaseError) {
-        console.error(error.message);
-      }
-      console.error(error);
-    }
-  };
-
   useEffect(() => {
-    getProject();
+    getProject(params, setProject);
   }, []);
 
   return (
